Use threads instead of the std library event loop
The event loop is very jank, inefficient and uses a lot of memory. Just
having a thread for each producer and a lock to share the entire state
performs a lot better as it uses not cpu when idle.

This is a very "to the point" port to threads. Not really clean or
anything. I think, idealy, I would just want it to be single threaded.